renamed the PostNotFound component and modifiied relevant files
