--- src/components/PostNotFound/PostNotFound.module.css
@@ -1,24 +1,0 @@
-.section {
-  position: relative;
-  margin: 20px auto;
-  width: 100%;
-  background-color: #fff;
-  padding: 4px;
-  border-radius: 0.5rem;
-}
-
-.error {
-  font-size: 1.2rem;
-  color: #575656;
-  font-weight: 500;
-}
-
-@media (min-width: 440px) and (max-width: 570px) {
-  .section {
-    margin-left: 50px;
-  }
-  .error {
-    font-size: 1.1rem;
-    font-style: oblique;
-  }
-}
